requirements.txt
# === MAVLink та автопілот ===
pymavlink>=2.4.30
dronekit>=2.9.2

# === Інтерфейс користувача ===
customtkinter>=5.2.1

# === Обробка конфігурацій ===
pyyaml>=6.0.1
python-dotenv>=1.0.1

# === Тестування ===
pytest>=8.1.1

# === Утиліти ===
typing_extensions>=4.10.0
